+++ packages/app-btc/src/utils/operationHelper.ts
@@ -2,7 +2,7 @@ import { ISDK } from '@cypherock/sdk-core';
 import { DeviceAppError, DeviceAppErrorType } from '@cypherock/sdk-interfaces';
 import { OnStatus } from '@cypherock/sdk-utils';
 import { DeepPartial, Exact, Query, Result } from '../proto/generated/btc/core';
-import { assertOrThrowInvalidResult, parseCoreError } from './asserts';
+import { assertOrThrowInvalidResult, parseCommonError } from './asserts';
 
 export function decodeResult(data: Uint8Array) {
   let result: Result;
@@ -10,7 +10,7 @@ export function decodeResult(data: Uint8Array) {
   try {
     result = Result.decode(data);
   } catch (error) {
-    throw new DeviceAppError(DeviceAppErrorType.INVALID_RESULT);
+    throw new DeviceAppError(DeviceAppErrorType.INVALID_MSG_FROM_DEVICE);
   }
 
   return result;
@@ -59,7 +59,7 @@ export class OperationHelper<Q extends QueryKey, R extends ResultKey> {
 
     const retrunObj = result[this.resultKey] as Result[R];
     assertOrThrowInvalidResult(retrunObj);
-    parseCoreError((result[this.resultKey] as any).coreError);
+    parseCommonError((result[this.resultKey] as any).commonError);
 
     return retrunObj;
   }