+++ packages/app-btc/tests/03.getXpubs/__fixtures__/invalidData.ts
@@ -34,7 +34,8 @@ const commonParams = {
     },
   ],
   errorInstance: DeviceAppError,
-  errorMessage: deviceAppErrorTypeDetails[DeviceAppErrorType.INVALID_RESULT],
+  errorMessage:
+    deviceAppErrorTypeDetails[DeviceAppErrorType.INVALID_MSG_FROM_DEVICE],
 };
 
 const invalidData: IGetXpubsTestCase[] = [