+++ packages/app-evm/tests/03.signTxn/__fixtures__/error.ts
@@ -58,7 +58,7 @@ const withInvalidResult: ISignTxnTestCase = {
   ],
   errorInstance: DeviceAppError,
   errorMessage:
-    deviceAppErrorTypeDetails[DeviceAppErrorType.INVALID_RESULT_ON_DEVICE],
+    deviceAppErrorTypeDetails[DeviceAppErrorType.INVALID_MSG_FROM_DEVICE],
 };
 
 const error: ISignTxnTestCase[] = [withUnknownError, withInvalidResult];