+++ packages/core/src/operations/raw/sendAbort.ts
@@ -41,11 +41,10 @@ export const sendAbort = async ({
   const usableConfig = config.v3;
 
   const packetsList = encodePacket({
-    data: '',
+    rawData: '',
     version,
     sequenceNumber,
     packetType: usableConfig.commands.PACKET_TYPE.ABORT,
-    isProto: false,
   });
 
   if (packetsList.length === 0) {