+++ apps/node/CHANGELOG.md
@@ -1,5 +1,16 @@
 # nodejs-test
 
+## 0.0.2
+
+### Patch Changes
+
+- d1e6255: Updated core msg structure
+- Updated dependencies [d1e6255]
+  - @cypherock/sdk-app-manager@0.0.2
+  - @cypherock/sdk-core@0.0.2
+  - @cypherock/sdk-hw-hid@0.0.2
+  - @cypherock/sdk-hw-serialport@0.0.2
+
 ## 0.0.1
 
 ### Patch Changes