+++ packages/core/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @cypherock/sdk-core
 
+## 0.0.2
+
+### Patch Changes
+
+- d1e6255: Updated core msg structure
+- Updated dependencies [d1e6255]
+  - @cypherock/sdk-interfaces@0.0.2
+  - @cypherock/sdk-utils@0.0.2
+
 ## 0.0.1
 
 ### Patch Changes