+++ packages/util-eslint-config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @cypherock/eslint-config
 
+## 0.0.2
+
+### Patch Changes
+
+- d1e6255: Updated core msg structure
+
 ## 0.0.1
 
 ### Patch Changes